requirements/requirements.txt
python-dotenv>=1.0.0
click>=8.1.7
requests>=2.31.0
rich>=13.7.0
pydantic>=2.5.3